Renamed and moved maven files/classes
